explainability: refactor after linting

+++ molpipeline/explainability/explanation.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import dataclasses
 
+import numpy as np
 import numpy.typing as npt
 
 from molpipeline.abstract_pipeline_elements.core import RDKitMol
@@ -14,17 +15,17 @@ class Explanation:
     """Class representing explanations of a prediction."""
 
     # input data
-    feature_vector: npt.NDArray[float] | None = None
+    feature_vector: npt.NDArray[np.float_] | None = None
     feature_names: list[str] | None = None
     molecule: RDKitMol | None = None
-    prediction: float | npt.NDArray[float] | None = None
+    prediction: float | npt.NDArray[np.float_] | None = None
 
     # explanation results mappable to the feature vector
-    feature_weights: npt.NDArray[float] | None = None
+    feature_weights: npt.NDArray[np.float_] | None = None
 
     # explanation results mappable to the molecule.
-    atom_weights: npt.NDArray[float] | None = None
-    bond_weights: npt.NDArray[float] | None = None
+    atom_weights: npt.NDArray[np.float_] | None = None
+    bond_weights: npt.NDArray[np.float_] | None = None
 
     def is_valid(self) -> bool:
         """Check if the explanation is valid.
